Add React Router and page components
