+++ src/App.tsx
@@ -1,45 +1,30 @@
-import { useEffect, useState } from "react";
-import { auth, provider } from "./firebase";
-import {
-  signInWithPopup,
-  signOut,
-  onAuthStateChanged,
-  type User,
-} from "firebase/auth";
+import { BrowserRouter, Routes, Route, Link } from 'react-router-dom';
+import Home from './pages/Home';
+import Login from './pages/Login';
+import Product from './pages/Product';
 
 function App() {
-  const [user, setUser] = useState<User | null>(null);
-
-  useEffect(() => {
-    const unsubscribe = onAuthStateChanged(auth, (currentUser) => {
-      setUser(currentUser);
-    });
-    return unsubscribe;
-  }, []);
-
-  const handleLogin = async () => {
-    try {
-      await signInWithPopup(auth, provider);
-    } catch (err) {
-      console.error(err);
-    }
-  };
-
-  const handleLogout = async () => {
-    await signOut(auth);
-  };
-
   return (
-    <div>
-      {user ? (
-        <>
-          <p>👋 Hello, {user.displayName}</p>
-          <button onClick={handleLogout}>登出</button>
-        </>
-      ) : (
-        <button onClick={handleLogin}>使用 Google 登入</button>
-      )}
-    </div>
+    <BrowserRouter>
+      <nav>
+        <ul>
+          <li>
+            <Link to="/">Home</Link>
+          </li>
+          <li>
+            <Link to="/login">Login</Link>
+          </li>
+          <li>
+            <Link to="/product">Product</Link>
+          </li>
+        </ul>
+      </nav>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="/login" element={<Login />} />
+        <Route path="/product" element={<Product />} />
+      </Routes>
+    </BrowserRouter>
   );
 }
