+++ src/pages/Home.tsx
@@ -1,0 +1,10 @@
+function Home() {
+  return (
+    <div>
+      <h2>Home</h2>
+      <p>Welcome to the shop!</p>
+    </div>
+  );
+}
+
+export default Home;
