+++ src/pages/Login.tsx
@@ -1,0 +1,46 @@
+import { useEffect, useState } from 'react';
+import { auth, provider } from '../firebase';
+import {
+  onAuthStateChanged,
+  signInWithPopup,
+  signOut,
+  type User,
+} from 'firebase/auth';
+
+function Login() {
+  const [user, setUser] = useState<User | null>(null);
+
+  useEffect(() => {
+    return onAuthStateChanged(auth, (currentUser) => {
+      setUser(currentUser);
+    });
+  }, []);
+
+  const handleLogin = async () => {
+    try {
+      await signInWithPopup(auth, provider);
+    } catch (err) {
+      console.error(err);
+    }
+  };
+
+  const handleLogout = async () => {
+    await signOut(auth);
+  };
+
+  return (
+    <div>
+      <h2>Login</h2>
+      {user ? (
+        <>
+          <p>👋 Hello, {user.displayName}</p>
+          <button onClick={handleLogout}>登出</button>
+        </>
+      ) : (
+        <button onClick={handleLogin}>使用 Google 登入</button>
+      )}
+    </div>
+  );
+}
+
+export default Login;
