+++ src/pages/Product.tsx
@@ -1,0 +1,10 @@
+function Product() {
+  return (
+    <div>
+      <h2>Product</h2>
+      <p>Details about our digital products will go here.</p>
+    </div>
+  );
+}
+
+export default Product;
